CloudFunctions/post_run_commentary_function/requirements.txt
# CloudFunctions/post_run_commentary_function/requirements.txt
functions-framework==3.*
google-generativeai==0.5.*
